changed sample directory, moved url_exists
