+++ sunpy/util/net.py
@@ -12,7 +12,7 @@ import sys
 import shutil
 
 # For Content-Disposition parsing
-from urllib2 import urlopen
+import urllib2
 import urlparse
 from email.parser import FeedParser
 from unicodedata import normalize
@@ -22,7 +22,7 @@ from sunpy.util import replacement_filename
 
 __all__ = ['slugify','get_content_disposition', 'get_filename',
            'get_system_filename', 'get_system_filename_slugify',
-           'download_file', 'download_fileobj', 'check_download_file']
+           'download_file', 'download_fileobj', 'check_download_file', 'url_exists']
 
 # Characters not allowed in slugified version.
 _punct_re = re.compile(r'[:\t !"#$%&\'()*\-/<=>?@\[\\\]^_`{|},.]+')
@@ -92,7 +92,7 @@ def download_file(url, directory, default=u'file', overwrite=False):
     Content-Disposition header, otherwise get from path of url. Fall
     back to default if both fail. Only overwrite existing files when
     overwrite is True. """
-    opn = urlopen(url)
+    opn = urllib2.urlopen(url)
     try:
         path = download_fileobj(opn, directory, url, default, overwrite)
     finally:
@@ -161,3 +161,33 @@ def check_download_file(filename, remotepath, download_dir, remotename=None,
 
         download_file(urlparse.urljoin(remotepath, remotename),
                       download_dir, default=filename, overwrite=replace)
+
+def url_exists(url, timeout=2):
+    """
+    Checks whether a url is online.
+
+    Parameters
+    ----------
+    url: str
+        A string containing a URL
+
+    Returns
+    -------
+    value: bool
+
+    Examples
+    --------
+    >>> from sunpy.net.helio import parser
+    >>> url_exists('http://www.google.com')
+    True
+    >>> url_exists('http://aslkfjasdlfkjwerf.com')
+    False
+    """
+    try:
+        urllib2.urlopen(url, timeout=timeout)
+    except urllib2.HTTPError, e:
+        return False
+    except urllib2.URLError, e:
+        return False
+    else:
+        return True
